pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aia-sdk"
version = "0.2.0"
description = "Official Python SDK for the Agent Identity and Capability Protocol (AICP)"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [
    { name = "Sealroot", email = "sdk@sealroot.com" }
]
keywords = ["ai", "agents", "identity", "security", "cryptography", "aicp"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security :: Cryptography",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "httpx>=0.25.0",
    "pynacl>=1.5.0",
]

[project.urls]
Homepage = "https://sealroot.com"
Documentation = "https://docs.sealroot.com/sdks/python"
Repository = "https://github.com/aia-protocol/sdk-python"
Issues = "https://github.com/aia-protocol/sdk-python/issues"
